server/worker: refactor worker into submodules
